chore: move posthog-node code to server dir
