+++ src/server/posthog/index.ts
@@ -1,2 +1,3 @@
 export * from "./identify";
 export * from "./flags";
+export * from "./client";
